Move parser interfaces to the correct generic namespace
