+++ src/backend/main/py.main/caliopen_main/interfaces/__init__.py
@@ -1,3 +1,12 @@
-import IO
-import storage
+# -*- coding: utf-8 -*-
+"""Caliopen main interfaces definitions."""
+from __future__ import absolute_import, print_function, unicode_literals
 
+from .IO import JsonDictIO, ProtobufIO, DictIO, JsonIO
+from .storage import DbIO, IndexIO
+
+from .parser import IAttachmentParser, IMessageParser, IParticipantParser
+
+__all__ = ['JsonDictIO', 'ProtobufIO', 'DictIO', 'JsonIO',
+           'DbIO', 'IndexIO', 'IAttachmentParser', 'IMessageParser',
+           'IParticipantParser']
